+++ src/utils/auxiliar_structures/queue.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stddef.h>
 /**
  * This is a highly efficient implementation of a queue. It has no safety checks for sizes, overflows, etc.
  * Therefore, it is the programmer's responsibility to use it correctly.
@@ -6,11 +7,11 @@
 
 typedef struct Queue Queue;
 
-Queue *createQueue(unsigned int size);
+Queue *createQueue(size_t size);
 
-void enqueue(Queue *q, unsigned int index);
+void enqueue(Queue *q, size_t index);
 
-unsigned int dequeue(Queue *q);
+size_t dequeue(Queue *q);
 
 bool is_queue_empty(const Queue *q);
